@@ -23,6 +23,12 @@
     "@types/react-router-dom": "^4.0.8",
     "@types/styled-components": "^3.0.1",
     "babel-plugin-styled-components": "^1.8.0",
+    "eslint": "^5.6.1",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-config-prettier": "^3.1.0",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-jsx-a11y": "^6.1.1",
+    "eslint-plugin-react": "^7.11.1",
     "gatsby-plugin-react-helmet": "^3.0.0",
     "gatsby-plugin-resolve-src": "^2.0.0-beta.1",
     "gatsby-plugin-sass": "^2.0.1",
@@ -44,12 +50,10 @@
     "stylelint-processor-styled-components": "^1.0.0",
     "ts-loader": "^3.1.0",
     "tslint": "^5.11.0",
-    "tslint-config-prettier": "^1.15.0",
-    "tslint-eslint-rules": "^5.4.0",
-    "tslint-plugin-prettier": "^1.3.0",
     "tslint-react": "3.2.0",
     "tslint-react-a11y": "^1.0.0",
-    "typescript": "^3.0.3"
+    "typescript": "^3.0.3",
+    "typescript-eslint-parser": "^20.0.0"
   },
   "dependencies": {
     "axios": "^0.15.3",
@@ -57,8 +61,8 @@
     "gatsby-image": "^2.0.0-rc.4",
     "lodash": "^4.17.4",
     "ramda": "^0.25.0",
-    "react": "^16.4.1",
-    "react-dom": "^16.4.1",
+    "react": "^16.7.0-alpha",
+    "react-dom": "^16.7.0-alpha",
     "react-helmet": "^5.0.1",
     "react-modal": "^3.5.1",
     "react-redux": "^5.0.6",
